[tensorflow/compiler/{jit/compilability_check_util.cc,jit/increase_dynamism_for_auto_jit_pass.cc,mlir/python/mlir.cc,xla/tests/array_elementwise_ops_test.cc}, tensorflow/{core/kernels/matmul_op_impl.h,lite/toco/graph_transformations/unroll_batch_matmul.cc}] Improve vector space reservations; use better size types

+++ tensorflow/compiler/mlir/python/mlir.cc
@@ -232,7 +232,7 @@ std::string ImportGraphDef(const std::string &proto,
     Set_TF_Status_from_Status(status, s);
     return "// error";
   }
-  for (unsigned i = 0, e = node_names.size(); i < e; i++) {
+  for (size_t i = 0, e = node_names.size(); i < e; i++) {
     const string &name = node_names[i];
     if (name.empty()) continue;
